mata: Add HBTP (touch) support.
This is using the services and libs from the official TWRP bringup. I would reccomend if you wanted to do things properly grabbing a factory image and extracting all the necessary services and libs, then relinking the services. It's a process I dont want to get into because it takes quite a bit of time. I'll put my attempt at it on a different branch.

Change-Id: If2e25a5cca03a0cde7f678061ec7c9de13ec3f52